Coral Code fixed

+++ small_bot/src/small_bot/CoralMain.py
@@ -116,7 +116,7 @@ class CoralMain:
                             cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
 
             cv2.circle(orig, centroid, 4, (0, 255, 0), 2)
-            cv2.circle(orig, (self.w/2, self.h/2), (0, 0, 255), 2)
+            cv2.circle(orig, (int(self.w/2), int(self.h/2)), (0, 0, 255), 2)
             if centroid is not None:
                 self.go_to_test(centroid)
